+++ packages/orq-rc/src/orq_ai_sdk/models/filelistop.py
@@ -74,7 +74,7 @@ class FileListData(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = parse_datetime("2025-05-21T12:58:37.403Z")
+    created: Optional[datetime] = parse_datetime("2025-05-22T05:50:29.352Z")
     r"""The date and time the resource was created"""