+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -125,7 +125,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01JVSESDPRJX6320YY0QT98EVE"
+        "01JVV8P6D0YGSD58G9N093J39J"
     )
     r"""The id of the resource"""